Fix bug where Nyquist frequency was getting a phase shift
This meant that the output timeseries would be complex, and the real
part would then be taken, which affects all of the frequency
components afterwards. Not shifting the Nyquist frequency means
the output timeseries is automatically real. Note that by design,
this bug was always encountered in the original version of the code,
because the time series was padded to even length